Moved common libraries to aureycore-foundation
