+++ .gitignore
@@ -7,6 +7,7 @@
 
 # REMOVE WHOLE .idea folder
 .idea
+/*.iml
 
 # User-specific stuff
 .idea/**/workspace.xml
