+++ src/main/java/com/tyrengard/aureycore/Aureycore.java
@@ -17,5 +17,6 @@ public class Aureycore extends JavaPlugin {
     public void onDisable() {
 //        ActionBarEngine.stop(false);
         DamageIndicatorAPI.stop();
+
     }
 }
