--- src/main/java/com/tyrengard/aureycore/common/armorevents/Constants.java
@@ -1,31 +1,0 @@
-package com.tyrengard.aureycore.common.armorevents;
-
-import org.bukkit.Material;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-import static org.bukkit.Material.*;
-
-class Constants {
-    static final List<Material> BLOCKED_MATERIALS = new ArrayList<>(Arrays.asList(
-            LEGACY_BED_BLOCK, LEGACY_FENCE_GATE, LEGACY_IRON_DOOR_BLOCK, LEGACY_IRON_DOOR, LEGACY_WOODEN_DOOR,
-            LEGACY_DIODE, LEGACY_DIODE_BLOCK_OFF, LEGACY_DIODE_BLOCK_ON, LEGACY_WOOD_BUTTON, LEGACY_TRAP_DOOR,
-            LEGACY_REDSTONE_COMPARATOR_OFF, LEGACY_REDSTONE_COMPARATOR_ON, LEGACY_FENCE, LEGACY_NETHER_FENCE,
-            LEGACY_SILVER_SHULKER_BOX, LEGACY_DAYLIGHT_DETECTOR_INVERTED,
-            FURNACE, CHEST, TRAPPED_CHEST, BEACON, DISPENSER, DROPPER, HOPPER, CRAFTING_TABLE, ENCHANTING_TABLE,
-            ENDER_CHEST, ANVIL, SPRUCE_FENCE_GATE, BIRCH_FENCE_GATE, ACACIA_FENCE_GATE, JUNGLE_FENCE_GATE,
-            DARK_OAK_FENCE_GATE, SPRUCE_DOOR, BIRCH_DOOR, JUNGLE_DOOR, ACACIA_DOOR, DARK_OAK_DOOR, STONE_BUTTON,
-            IRON_TRAPDOOR, SPRUCE_FENCE, BIRCH_FENCE, JUNGLE_FENCE, DARK_OAK_FENCE, ACACIA_FENCE, BREWING_STAND,
-            CAULDRON, LEGACY_SIGN_POST, LEGACY_WALL_SIGN, LEGACY_SIGN, ACACIA_SIGN, ACACIA_WALL_SIGN, BIRCH_SIGN,
-            BIRCH_WALL_SIGN, DARK_OAK_SIGN, DARK_OAK_WALL_SIGN, JUNGLE_SIGN, JUNGLE_WALL_SIGN, OAK_SIGN, OAK_WALL_SIGN,
-            SPRUCE_SIGN, SPRUCE_WALL_SIGN, LEVER, DAYLIGHT_DETECTOR, BARREL, BLAST_FURNACE, SMOKER, CARTOGRAPHY_TABLE,
-            COMPOSTER, GRINDSTONE, LECTERN, LOOM, STONECUTTER, BELL,
-
-            // Shulker boxes
-            BLACK_SHULKER_BOX, BLUE_SHULKER_BOX, BROWN_SHULKER_BOX, CYAN_SHULKER_BOX, GRAY_SHULKER_BOX,
-            GREEN_SHULKER_BOX, LIGHT_BLUE_SHULKER_BOX, LIME_SHULKER_BOX, MAGENTA_SHULKER_BOX, ORANGE_SHULKER_BOX,
-            PINK_SHULKER_BOX, PURPLE_SHULKER_BOX, RED_SHULKER_BOX, WHITE_SHULKER_BOX, YELLOW_SHULKER_BOX
-    ));
-}
