--- src/main/java/com/tyrengard/aureycore/common/armorevents/PlayerArmorChangeEvent.java
@@ -1,4 +1,0 @@
-package com.tyrengard.aureycore.common.armorevents;
-
-public class PlayerArmorChangeEvent {
-}
