--- src/main/java/com/tyrengard/aureycore/common/interfaces/TriConsumer.java
@@ -1,6 +1,0 @@
-package com.tyrengard.aureycore.common.interfaces;
-
-@FunctionalInterface
-public interface TriConsumer<A, B, C> {
-	void accept(A a, B b, C c);
-}
