--- src/main/java/com/tyrengard/aureycore/common/struct/UUIDDataType.java
@@ -1,36 +1,0 @@
-package com.tyrengard.aureycore.common.struct;
-
-import java.nio.ByteBuffer;
-import java.util.UUID;
-
-import org.bukkit.persistence.PersistentDataAdapterContext;
-import org.bukkit.persistence.PersistentDataType;
-
-public class UUIDDataType implements PersistentDataType<byte[], UUID> {
-	public final static UUIDDataType instance = new UUIDDataType();
-	@Override
-    public Class<byte[]> getPrimitiveType() {
-        return byte[].class;
-    }
-
-    @Override
-    public Class<UUID> getComplexType() {
-        return UUID.class;
-    }
-
-    @Override
-    public byte[] toPrimitive(UUID complex, PersistentDataAdapterContext context) {
-        ByteBuffer bb = ByteBuffer.wrap(new byte[16]);
-        bb.putLong(complex.getMostSignificantBits());
-        bb.putLong(complex.getLeastSignificantBits());
-        return bb.array();
-    }
-
-    @Override
-    public UUID fromPrimitive(byte[] primitive, PersistentDataAdapterContext context) {
-        ByteBuffer bb = ByteBuffer.wrap(primitive);
-        long firstLong = bb.getLong();
-        long secondLong = bb.getLong();
-        return new UUID(firstLong, secondLong);
-    }
-}
