--- src/main/java/com/tyrengard/aureycore/common/utils/NumberUtils.java
@@ -1,11 +1,0 @@
-package com.tyrengard.aureycore.common.utils;
-
-public class NumberUtils {
-	public static int getRandomInt(int from, int to) {
-		return (int)(Math.random() * to + from);
-	}
-	
-	public static double roundToTwoDecimalPlaces(double d) {
-		return (double) Math.round(d * 100) / 100;
-	}
-}
