--- src/main/java/com/tyrengard/aureycore/common/utils/PlayerUtils.java
@@ -1,16 +1,0 @@
-package com.tyrengard.aureycore.common.utils;
-
-import org.bukkit.Bukkit;
-import org.bukkit.OfflinePlayer;
-
-public class PlayerUtils {
-    public static OfflinePlayer getOfflinePlayer(String name) {
-        for (OfflinePlayer player : Bukkit.getOfflinePlayers()) {
-            String playerName = player.getName();
-            if (name.equalsIgnoreCase(playerName))
-                return player;
-        }
-
-        return null;
-    }
-}
