--- src/main/java/com/tyrengard/aureycore/common/utils/PotionEffectUtils.java
@@ -1,5 +1,0 @@
-package com.tyrengard.aureycore.common.utils;
-
-public class PotionEffectUtils {
-	
-}
