+++ src/main/java/com/tyrengard/aureycore/customguis/Information.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import com.tyrengard.aureycore.common.utils.ItemUtils;
+import com.tyrengard.aureycore.foundation.common.utils.ItemUtils;
 import org.bukkit.Material;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemFlag;
