+++ src/main/java/module-info.java
@@ -2,14 +2,10 @@ module com.tyrengard.aureycore {
     requires org.bukkit;
     requires java.logging;
     requires org.jetbrains.annotations;
-    exports com.tyrengard.aureycore.common.anvilevents;
-    exports com.tyrengard.aureycore.common.stringformat;
-    exports com.tyrengard.aureycore.common.interfaces;
-    exports com.tyrengard.aureycore.common.struct;
-    exports com.tyrengard.aureycore.common.utils;
+    requires com.tyrengard.aureycore.foundation;
+
     exports com.tyrengard.aureycore.playerstate;
     exports com.tyrengard.aureycore.damageindicator;
     exports com.tyrengard.aureycore.customguis;
     exports com.tyrengard.aureycore;
-    exports com.tyrengard.aureycore.common.armorevents;
 }
